+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/config/S3Config.java
@@ -1,4 +1,4 @@
-package com.startingblue.fourtooncookie.aws.s3;
+package com.startingblue.fourtooncookie.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
