+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/diary/service/DiaryS3Service.java
@@ -1,6 +1,6 @@
 package com.startingblue.fourtooncookie.diary.service;
 
-import com.startingblue.fourtooncookie.aws.s3.service.S3Service;
+import com.startingblue.fourtooncookie.aws.s3.S3Service;
 import com.startingblue.fourtooncookie.global.converter.image.ImageConverter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
